samples/cosmosdb-nosql/requirements.txt
langchain-azure-cosmosdb
langchain-openai
langgraph
python-dotenv
